api/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
google-cloud-firestore==2.16.0
firebase-admin==6.5.0
pydantic==2.8.2
pydantic-settings==2.4.0
sympy==1.12
httpx==0.27.0
python-dotenv==1.0.1
ruff==0.5.4
black==24.4.2
jsonschema==4.23.0
pytest==8.3.2
google-cloud-aiplatform==1.64.0
